Revert "User session created."
This reverts commit fb777d5657db245b57ed809520ff98c30a585aa7.
