--- src/main/java/com/Repository/UserSessionRepository.java
@@ -1,9 +1,0 @@
-package com.Repository;
-
-import com.Model.User;
-import com.Model.UserSession;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserSessionRepository extends CrudRepository<UserSession, Long> {
-    public UserSession findBySessionId (String sessionId);
-}
