+++ src/main/resources/static/css/general.css
@@ -7,8 +7,3 @@ table th td{
     position: relative;
     top: 15px;
 }
-.toprightcorner {
-    position:absolute;
-    top:30px;
-    right:30px;
-}
